market-digest-starter/requirements.txt
feedparser
beautifulsoup4
requests
pandas
numpy
python-dateutil
pytz
yfinance
Pillow
matplotlib
rapidfuzz
